Move collate functions to util
