+++ src/netspresso_trainer/dataloaders/builder.py
@@ -7,10 +7,9 @@ import torch.distributed as dist
 from loguru import logger
 
 from .augmentation.registry import TRANSFORM_DICT
-from .classification import classification_mix_collate_fn, classification_onehot_collate_fn
-from .detection import detection_collate_fn
 from .registry import CREATE_TRANSFORM, CUSTOM_DATASET, DATA_SAMPLER, HUGGINGFACE_DATASET
 from .utils.loader import create_loader
+from .utils.collate_fn import classification_mix_collate_fn, classification_onehot_collate_fn, detection_collate_fn
 
 TRAIN_VALID_SPLIT_RATIO = 0.9
